fix GCS types to work. We don't export a constructor, instead we export a namespace + function combo that works
